requirements.txt
pillow
requests
openpyxl
playwright
requests